Tested list,remove, add and find flags as well as fixed some minor bugs.
Also added some reserve keywords list which can not be used for naming the host or website.

added code for some edge cases as well.

+++ libs/remove.js
@@ -4,20 +4,23 @@ const { input, over } = require("../util/input.js");
 const fs = require("fs");
 
 const remove = async () => {
-
+    if(process.argv[3] === undefined){
+        console.log("One more parameter required,")
+        console.log("Try passdom -r <Website name> or passdom remove <Website name> instead.")
+        return over()
+    }
 
     if (Object.keys(data).includes(process.argv[3].toLowerCase())) {
 
-        let inputVal = await input("Are you sure you want to remove this account? (y/n) ")
+        let inputVal = await input(`Are you sure you want to remove credentials for "${process.argv[3]}"? (y/n) `)
         let val = inputVal.toLowerCase();
         if (val === "y" || val === "yes") {
             delete data[process.argv[3].toLowerCase()]
             fs.writeFile(`${__dirname}/../localStorage/cred.json`, JSON.stringify(data, null, 2), err => {
                 if (err) {
-                    console.log('Error removing data', err.message)
-                } else {
-                    console.log(`Removed the data`)
-                }
+                    console.log('Error removing credentials', err.message)
+                } 
+                console.log(`Successfully removed credentials for ${process.argv[3].toLowerCase()}`) 
             })
             over()
         } else {
@@ -25,7 +28,7 @@ const remove = async () => {
         }
     }
     else {
-        console.log("Data not found, Hence not removed!");
+        console.log(`Credentials for ${process.argv[3].toLowerCase()} not found, Hence not removed!`);
         over();
     }
 }
